backend/requirements.txt
sqlmodel
sqlalchemy
pytest
ruff
fastapi
minio
python-multipart  # needed by FastAPI for UploadFile / multipart/form-data
python-dotenv
uvicorn
httpx # needed by TestClient
psycopg[binary] # needed for psql connection
pika
pillow

--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.6.0
transformers>=4.45
numpy<2